refactor: isolating creation methods in repositories
