+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
 var databaseSetup = new DatabaseSetup(databaseConfig);
@@ -31,18 +32,9 @@ if (modelName == "Computer")
         var ram = args[3];
         var processor = args [4]; 
         
-        var connection = new SqliteConnection(databaseConfig.ConnectionString);
-        connection.Open();
+        var computer = new Computer(id, ram, processor);
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);"; 
-        command.Parameters.AddWithValue("$id",id);
-        command.Parameters.AddWithValue("$ram",ram);
-        command.Parameters.AddWithValue("$processor",processor);
-
-            
-        command.ExecuteNonQuery();
-        connection.Close();
+        computerRepository.Save(computer);
 
     }
 }
@@ -71,20 +63,9 @@ if (modelName == "Lab")
         var name = args [4]; 
         var block = args [5];
         
-        var connection = new SqliteConnection(databaseConfig.ConnectionString);
-        connection.Open();
+       var lab = new Lab(id, number, name, block);
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Labs VALUES($id, $number, $name, $block)"; 
-
-        command.Parameters.AddWithValue("$id",id);
-        command.Parameters.AddWithValue("$number",number);
-        command.Parameters.AddWithValue("name", name);
-        command.Parameters.AddWithValue("$block",block);
-
-            
-        command.ExecuteNonQuery();
-        connection.Close();
+       labRepository.Save(lab);
 
     }
 }
